Turn gds_send_request_t and gds_wait_request_t to opaque objects
